bugfix - now Python can load huge signals such as RC

+++ Internal/MedPyExport/MedPyExport/MPDictionary.cpp
@@ -21,7 +21,7 @@ int MPDictionary::id(string& signame) { return o->dict.id(signame); }
 
 string MPDictionary::name(int id) { return o->dict.name(id); };
 
-void MPDictionary::prep_sets_lookup_table(int section_id, const std::vector<string>& set_names, MEDPY_NP_OUTPUT(char** lut_array, int* lut_size)) {
+void MPDictionary::prep_sets_lookup_table(int section_id, const std::vector<string>& set_names, MEDPY_NP_OUTPUT(char** lut_array, unsigned long long* lut_size)) {
 	std::vector<char> lut;
 	int retval = o->dict.prep_sets_lookup_table(section_id, set_names, lut);
 	if (retval != 0) {
@@ -29,11 +29,11 @@ void MPDictionary::prep_sets_lookup_table(int section_id, const std::vector<stri
 	}
 	
 	*lut_array = (char*)malloc(sizeof(char)*lut.size());
-	*lut_size = (int)lut.size();
+	*lut_size = lut.size();
 	memcpy(*lut_array, lut.data(), lut.size());
 }
 
-MPIntVecIntMapAdaptor MPDictionary::get_members_to_all_sets(int section_id, MEDPY_NP_INPUT(int* members_array, int members_size)) {
+MPIntVecIntMapAdaptor MPDictionary::get_members_to_all_sets(int section_id, MEDPY_NP_INPUT(int* members_array, unsigned long long members_size)) {
 	vector<int> members;
 	buf_to_vector(members_array, members_size, members);
 	unordered_map<int, std::vector<int>> Member2AllSets;
