+++ Internal/MedPyExport/MedPyExport/MPGlobal.h
@@ -27,7 +27,7 @@ public:
 
 class MPCommonLib {
 public:
-	static void print_auc_performance(MPSamples &samples, MEDPY_NP_INPUT(int* folds, int num_folds), string outFile);
+	static void print_auc_performance(MPSamples &samples, MEDPY_NP_INPUT(int* folds, unsigned long long num_folds), string outFile);
 	static void shuffleMatrix(MPFeatures& matrix);
 	static void read_predictor_from_file(MPPredictor& pred, string predictorFile);
 };
